Added a set of examples of using ColdSpring AOP in combination with Taffy / taffy:aopbean.  One example demonstrates wrapping caching around a resource, the second demonstrates how a developer could add a system-wide return format without having to add the code to each resource cfc.
There was one issue getting the formatting example to work, which was that I had to flip the access mode of representationOf() to public.  If we can find a way around that, I'm happy to revert it back to private.

Additionally, a change was made to core/api.cfc in the reload logic.  Because in all likelihood ColdSpring will be instantiated as application scope singleton, reload() as existed would not update changes to the ColdSpring config file.  Now, it also re-fires applicationStartEvent() so it actually resets ColdSpring (and user's custom application start events) as well.