Change to add user friendly documentation
Added a new docs.cfm to show user friendly and printable documentation
for users. The docs.cfm file will show when the dashboard is disabled,
and if the user has set the showDocsWhenDashboardDisabled to True.
showDocsWhenDashboardDisabled defaults to false in order to maintain
backward compatibility.

New settings:
showDocsWhenDashboardDisabled: True or False
docs.APIName: The name of the API to show in the documentation
docs.APIVersion: The version of the API to show in the documentation