Change resourceUri to a text input
When using longer resource links, the resourceUri <code> block flows
outside of the runner div and lays over the top of the docs. I’ve
changed the resourceUri to be a text input, monospace formatted, with a
select() onclick to make it easier to read the URL, and really simple
to select it and run it in a browser.