Exception for QoQ with blank column name
There are instances where a column name from a QoQ could return an
empty string as a column name, causing an exception when trying to
reference that column name when converting a query to an array. This
change will ensure that we have a valid column name before attempting
to reference it.