https://github.com/atuttle/Taffy/issues/239
Added the ability to specify a comma delimited list of URIs for taffy_uri, allowing you to specify multiple endpoint URIs to be served by the same resource .  It will not recognize commas inside of regular expression brackets {}.

Fixed an issue where status.skippedResources was being overwritten when hot-swapping.

Added tests for the new alias', as well as reconfiguring the tests to keep testing the skippedResources properly.

Added a test for conflicting URI errors.