Update docs.cfm to correct bad collapse behavior
On initial load, clicking the heading for a resource will attempt to expand the panel, when it is already expanded. A second click will then collapse the panel, and behavior will return to normal.

Added `class="in"` to the resource section of each panel so that the initial click on a heading will properly collapse the panel.